pyproject.toml
[project]
name = "score-reading"
version = "0.1.0"
description = "朗读/背诵口语评分 CLI 框架"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your@email.com"}
]
keywords = ["speech", "pronunciation", "scoring", "cli"]

dependencies = [
    "typer>=0.9.0",
    "pydub>=0.25.1",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
    "numpy>=1.24.0",
    "httpx>=0.25.0",
    "textgrid>=1.5",
    "rich>=13.0.0",
    "librosa>=0.10.0",
    "transformers>=4.30.0",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[project.scripts]
score_reading = "src.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true
